数据结构

+++ search/bianrySearch.cpp
@@ -1,6 +1,6 @@
 /**
  * @file: bianrySearch.cpp
- * @description: 
+ * @description:
  * @author: koritafei
  * @create: 2021-05-31 13:49
  * @version v0.1
@@ -9,11 +9,11 @@
 #include <iostream>
 #include <vector>
 
-template<class T>
-int bianrySearch(std::vector<T> &nums, int key) {
-  int left = 0, right = nums.size() - 1;
+template <class T>
+unsigned long bianrySearch(std::vector<T> &nums, int key) {
+  unsigned long left = 0, right = nums.size() - 1;
   while (left <= right) {
-    int middle = left + (right - left) / 2;
+    unsigned long middle = left + (right - left) / 2;
     if (nums[middle] == key) {
       return middle;
     } else if (nums[middle] < key) {
@@ -26,12 +26,12 @@ int bianrySearch(std::vector<T> &nums, int key) {
   return -1;
 }
 
-template<class T>
-int floor(std::vector<T> &nums, int key) {
+template <class T>
+unsigned long floor(std::vector<T> &nums, int key) {
   // 最大比key小的值
-  int left = 0, right = nums.size() - 1;
+  unsigned long left = 0, right = nums.size() - 1;
   while (left <= right) {
-    int middle = left + (right - left) / 2;
+    unsigned long middle = left + (right - left) / 2;
     if (nums[middle] == key) {
       return middle;
     } else if (nums[middle] > key) {
@@ -44,16 +44,16 @@ int floor(std::vector<T> &nums, int key) {
   return right;
 }
 
-template<class T>
-int ceil(std::vector<T> &nums, int key) {
+template <class T>
+unsigned long ceil(std::vector<T> &nums, int key) {
   // 最小的比key大的值
-  int left = 0, right = nums.size() - 1;
+  unsigned long left = 0, right = nums.size() - 1;
   while (left <= right) {
-    int middle = left + (right - left) / 2;
+    unsigned long middle = left + (right - left) / 2;
     if (nums[middle] == key) {
       return middle;
     } else if (nums[middle] > key) {
-      right = middle -1;
+      right = middle - 1;
     } else if (nums[middle] < key) {
       left = middle + 1;
     }
@@ -62,10 +62,9 @@ int ceil(std::vector<T> &nums, int key) {
   return left;
 }
 
-
 int main(int argc, char **argv) {
   std::vector<int> nums = {1, 4, 6, 7, 9, 10, 15, 1100};
-  int key = 1100;
+  int              key  = 1100;
   std::cout << bianrySearch(nums, key) << std::endl;
   key = 8;
   std::cout << bianrySearch(nums, key) << std::endl;
